fixed length of D+ and D- Traces, case progress.
reorganized traces and optimized the board layout. Make no mistake, it still looks bad. but at least its not as bad as it was.

I also made progress on the case, adding a hole to route cables from the side compartments into the main body.

note that fpr some reason the measurements for the inside are all screwed up, and a lot of things arent at 90 degree angles. I hope to fix that soon. for now though, any discrepancies in length are usually within 0.1mm, so it should be fine.