adjusted clearances in model, updated docs
changed thickness of deck to take advantage of a reduced carrier board size (I'll need to mod it tho), and added the embedded nuts to the rail system.

also updated the documentation to specify materials and the various considerations taken into account.

Linked to resources detailing precise dimensions for the metal components used in the deck.